* remove circular dependencies in step-generation by refactoring utils dir
