+++ protocol-designer/src/step-generation/getNextRobotStateAndWarnings/dispenseUpdateLiquidState.js
@@ -8,7 +8,7 @@ import {
   mergeLiquid,
   getWellsForTips,
   getLocationTotalVolume,
-} from '../utils'
+} from '../utils/misc'
 import type {
   RobotState,
   InvariantContext,
